Commenting and Adjustments
I spent a bit of time looking at the transitions and thinking of additive things I could add rather than the way that the transitions take place because I actually quite like the default. So far I've added some arcs that grow depending on how far the letter is spinning, and one of the circles shrinks as well.
Next is the ReadME and finishing the interpolation.